docs: update code
