+++ src/pages/pageHome/modal/addNewCat.tsx
@@ -1,27 +1,25 @@
-import { Cat } from 'components/main/main';
 import React from 'react';
+import { Cat } from 'types/models';
 import { Form } from '../form/form';
 import classes from './addNewCat.module.css';
 
-type PropsFunc = {
+type AddNewCatProps = {
   createCat: (cat: Cat) => void;
-  toggleModal: () => void;
+  toggleModal: (bool: boolean) => void;
 };
 
-export class AddNewCat extends React.Component<PropsFunc> {
+export class AddNewCat extends React.Component<AddNewCatProps> {
   handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     const form = e.target as HTMLFormElement;
     const formData = Object.fromEntries(new FormData(form).entries());
 
-    const randomImage = async () => {
-      return await fetch('https://source.unsplash.com/random?cat').then((data) => data.url);
-    };
+    const randomImage = await fetch('https://source.unsplash.com/random?cat').then((data) => data.url);
 
     const cat: Cat = {
       id: String(Math.random()),
       length: String(formData.length),
-      image: String(await randomImage()),
+      image: String(randomImage),
       minWeight: Number(formData.minWeight),
       maxWeight: Number(formData.maxWeight),
       minLife: Number(formData.minLife),
@@ -32,11 +30,12 @@ export class AddNewCat extends React.Component<PropsFunc> {
     this.props.createCat(cat);
     form.reset();
   };
+
   render() {
     return (
       <div className={classes.content__modal}>
         <div className={classes['wrapper_close-modal']}>
-          <button className={classes['close-modal']} onClick={this.props.toggleModal}>
+          <button className={classes['close-modal']} onClick={() => this.props.toggleModal(false)}>
             &times;
           </button>
         </div>
