+++ src/pages/pageHome/pageHome.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Card } from './card/card';
 import { Search } from './search/search';
 import classes from './pageHome.module.css';
-import { Cats } from 'components/main/main';
+import { Cats } from 'types/models';
 
 export class Home extends React.Component<Cats> {
   render() {
